Catalogue add and remove items on collision
Items can now be transferred from the master catalogue that contains all possible items to the player's catalogue which is initially empty.

Multiple items of the same type can be added.

Items are added when a rigid body enters the trigger box collider of the ground

items are removed when the same rigid body leaves the trigger box collider of the ground but not when it is deleted

To be created: have a button that deletes all items of 1 type that simultaneously deletes all items of 1 type in the catalogue